refactor: move test to datashare-tasks module #1653
