+++ datashare-tasks/src/test/java/org/icij/datashare/asynctasks/TaskManagerMemoryWithRedisTest.java
@@ -1,23 +1,15 @@
-package org.icij.datashare.tasks;
+package org.icij.datashare.asynctasks;
 
 import org.icij.datashare.PropertiesProvider;
-import org.icij.datashare.asynctasks.Task;
-import org.icij.datashare.asynctasks.TaskManagerMemory;
-import org.icij.datashare.asynctasks.TaskRepositoryRedis;
-import org.icij.datashare.asynctasks.TaskResult;
-import org.icij.datashare.db.JooqTaskRepository;
-import org.icij.datashare.db.RepositoryFactoryImpl;
 import org.icij.datashare.user.User;
 import org.icij.extract.redis.RedissonClientFactory;
 import org.icij.task.Options;
-import org.jooq.SQLDialect;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
 import org.redisson.api.RedissonClient;
 
-import javax.sql.DataSource;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
@@ -26,9 +18,9 @@ import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.fest.assertions.Assertions.assertThat;
 
 public class TaskManagerMemoryWithRedisTest {
-    @Mock DatashareTaskFactory factory;
+    TaskFactory factory = new TestFactory();
 
-    private org.icij.datashare.asynctasks.TaskManagerMemory taskManager;
+    private TaskManagerMemory taskManager;
     private final CountDownLatch waitForLoop = new CountDownLatch(1);
 
     @Before
@@ -45,16 +37,17 @@ public class TaskManagerMemoryWithRedisTest {
 
     @Test
     public void test_run_task() throws Exception {
-        String tid = taskManager.startTask(TestTask.class, User.local(), new HashMap<>());
+        String tid = taskManager.startTask(TestFactory.Sleep.class, User.local(), Map.of("duration", 10));
+
         taskManager.waitTasksToBeDone(1, SECONDS);
 
-        assertThat(taskManager.getTask(tid).getState()).isEqualTo(Task.State.ERROR);
-        assertThat(taskManager.getTask(tid).getError().getMessage()).isEqualTo("Cannot invoke \"Object.getClass()\" because \"factory\" is null");
+        assertThat(taskManager.getTask(tid).getState()).isEqualTo(Task.State.DONE);
+        assertThat(taskManager.getTask(tid).getResult()).isEqualTo(new TaskResult<>(10));
     }
 
     @After
     public void tearDown() throws Exception {
+        taskManager.clear();
         taskManager.close();
     }
-
 }
